+++ src/pages/exercises/SmartCam/components/index.ts
@@ -1,0 +1,4 @@
+export * from './CameraCanvas';
+export * from './CameraControls';
+export * from './CameraSection';
+export * from './IntroSection';